--- archive/v5/app/SushiClass.js
@@ -1,17 +1,0 @@
-var easeljs       = require( 'easeljs' ),
-	tweenjs       = require( 'tweenjs' ),
-	soundjs       = require( 'soundjs' ),
-	preloadjs     = require( 'preloadjs' ),
-	$             = require( 'jquery' ),
-
-	SushiClass  = function () {
-		createjs.Bitmap.prototype.constructor.apply( this,arguments );
-	};
-	
-	createjs.extend(SushiClass, createjs.Bitmap);
-	
-	SushiClass.prototype.tick = function( event ) {
-		this.x -= this.vX;
-	};
-
-module.exports = SushiClass;
