+++ archive/v5/app/music_notes.js
@@ -1,0 +1,149 @@
+
+module.exports = function Music() {
+	this.notes = [
+		4285.714285714285,
+		6428.5714285714275,
+		6857.142857142856,
+		9000,
+		9642.857142857143,
+		9857.142857142857,
+		11142.857142857143,
+		11357.142857142857,
+		11571.42857142857,
+		13714.285714285714,
+		15857.142857142859,
+		16071.428571428572,
+		16285.714285714286,
+		16714.285714285714,
+		19285.714285714283,
+		20571.42857142857,
+		22714.28571428571,
+		23142.857142857138,
+		25285.71428571428,
+		25928.57142857142,
+		26142.857142857134,
+		27428.57142857142,
+		27642.857142857134,
+		27857.142857142848,
+		29999.999999999993,
+		31285.714285714275,
+		31928.571428571417,
+		32142.85714285713,
+		32571.428571428558,
+		33857.14285714284,
+		34071.42857142856,
+		34285.714285714275,
+		35142.85714285713,
+		36428.57142857141,
+		37285.71428571427,
+		37714.285714285696,
+		39428.571428571406,
+		39857.14285714283,
+		40285.71428571426,
+		40499.99999999998,
+		41571.42857142855,
+		42428.571428571406,
+		42642.85714285712,
+		42857.14285714284,
+		43285.71428571427,
+		43714.285714285696,
+		44142.85714285712,
+		44571.42857142855,
+		44999.99999999998,
+		45857.14285714284,
+		47142.85714285712,
+		47571.42857142855,
+		47999.99999999998,
+		48428.571428571406,
+		49285.71428571426,
+		49499.99999999998,
+		49714.285714285696,
+		50357.14285714284,
+		50999.999999999985,
+		51214.2857142857,
+		53142.85714285713,
+		53571.42857142856,
+		53999.999999999985,
+		54214.2857142857,
+		54642.85714285713,
+		54857.14285714285,
+		54964.2857142857,
+		55285.714285714275,
+		56142.85714285713,
+		56999.999999999985,
+		57857.14285714285,
+		57964.2857142857,
+		58071.42857142856,
+		58178.57142857141,
+		58499.999999999985,
+		58714.2857142857,
+		59785.714285714275,
+		59892.85714285713,
+		59999.999999999985,
+		60214.285714285696,
+		60428.571428571406,
+		60857.14285714284,
+		61714.285714285696,
+		61821.42857142855,
+		62357.14285714283,
+		62464.28571428569,
+		62785.71428571426,
+		63428.571428571406,
+		63535.71428571426,
+		64071.42857142854,
+		64178.5714285714,
+		64499.99999999997,
+		64607.142857142826,
+		65035.71428571425,
+		65142.85714285711,
+		65249.99999999996,
+		66642.85714285712,
+		66857.14285714283,
+		67071.42857142854,
+		67714.28571428568,
+		68357.14285714283,
+		68571.42857142854,
+		68678.57142857139,
+		70071.42857142854,
+		70285.71428571425,
+		70499.99999999996,
+		71142.8571428571,
+		71999.99999999997,
+		72107.14285714283,
+		72214.28571428568,
+		72857.14285714283,
+		73071.42857142854,
+		73499.99999999997,
+		74357.14285714284,
+		74785.71428571428,
+		75214.28571428571,
+		76714.28571428571,
+		76928.57142857142,
+		77571.42857142857,
+		77785.71428571428,
+		78642.85714285714,
+		78750,
+		78857.14285714286,
+		78964.28571428571,
+		79285.71428571428,
+		79499.99999999999,
+		79928.57142857142,
+		80142.85714285713,
+		80249.99999999999,
+		80357.14285714284,
+		80999.99999999999,
+		81428.57142857142,
+		81642.85714285713,
+		82071.42857142857,
+		83357.14285714286,
+		83785.71428571429,
+		84000,
+		84107.14285714286,
+		85500,
+		85714.28571428571,
+		85928.57142857142,
+		86571.42857142857,
+		87428.57142857143,
+		87535.71428571429
+	];
+};
