+++ archive/v5/app/sushi_class.js
@@ -1,0 +1,17 @@
+var easeljs       = require( 'easeljs' ),
+	tweenjs       = require( 'tweenjs' ),
+	soundjs       = require( 'soundjs' ),
+	preloadjs     = require( 'preloadjs' ),
+	$             = require( 'jquery' );
+
+var SushiClass  = function () {
+	createjs.Bitmap.prototype.constructor.apply( this,arguments );
+};
+	
+createjs.extend(SushiClass, createjs.Bitmap);
+	
+SushiClass.prototype.tick = function() {
+	this.x -= this.vX;
+};
+
+module.exports = SushiClass;
