--- archive/v5/npm-debug.log
@@ -1,110 +1,0 @@
-0 info it worked if it ends with ok
-1 verbose cli [ '/home/s.yamagishi/.nvm/versions/node/v6.1.0/bin/node',
-1 verbose cli   '/home/s.yamagishi/.nvm/versions/node/v6.1.0/bin/npm',
-1 verbose cli   'install',
-1 verbose cli   'uglify-css',
-1 verbose cli   '--save-dev' ]
-2 info using npm@3.8.6
-3 info using node@v6.1.0
-4 silly loadCurrentTree Starting
-5 silly install loadCurrentTree
-6 silly install readLocalPackageData
-7 silly fetchPackageMetaData uglify-css
-8 silly fetchNamedPackageData uglify-css
-9 silly mapToRegistry name uglify-css
-10 silly mapToRegistry using default registry
-11 silly mapToRegistry registry https://registry.npmjs.org/
-12 silly mapToRegistry data Result {
-12 silly mapToRegistry   raw: 'uglify-css',
-12 silly mapToRegistry   scope: null,
-12 silly mapToRegistry   name: 'uglify-css',
-12 silly mapToRegistry   rawSpec: '',
-12 silly mapToRegistry   spec: 'latest',
-12 silly mapToRegistry   type: 'tag' }
-13 silly mapToRegistry uri https://registry.npmjs.org/uglify-css
-14 verbose request uri https://registry.npmjs.org/uglify-css
-15 verbose request no auth needed
-16 info attempt registry request try #1 at 16:42:16
-17 verbose request id c44d6870f27b2ae2
-18 http request GET https://registry.npmjs.org/uglify-css
-19 http 404 https://registry.npmjs.org/uglify-css
-20 verbose headers { 'content-type': 'application/json',
-20 verbose headers   'cache-control': 'max-age=0',
-20 verbose headers   'content-length': '2',
-20 verbose headers   'accept-ranges': 'bytes',
-20 verbose headers   date: 'Mon, 23 May 2016 07:42:18 GMT',
-20 verbose headers   via: '1.1 varnish',
-20 verbose headers   age: '0',
-20 verbose headers   connection: 'keep-alive',
-20 verbose headers   'x-served-by': 'cache-itm7426-ITM',
-20 verbose headers   'x-cache': 'MISS',
-20 verbose headers   'x-cache-hits': '0',
-20 verbose headers   'x-timer': 'S1463989338.052615,VS0,VE153',
-20 verbose headers   vary: 'Accept-Encoding' }
-21 silly get cb [ 404,
-21 silly get   { 'content-type': 'application/json',
-21 silly get     'cache-control': 'max-age=0',
-21 silly get     'content-length': '2',
-21 silly get     'accept-ranges': 'bytes',
-21 silly get     date: 'Mon, 23 May 2016 07:42:18 GMT',
-21 silly get     via: '1.1 varnish',
-21 silly get     age: '0',
-21 silly get     connection: 'keep-alive',
-21 silly get     'x-served-by': 'cache-itm7426-ITM',
-21 silly get     'x-cache': 'MISS',
-21 silly get     'x-cache-hits': '0',
-21 silly get     'x-timer': 'S1463989338.052615,VS0,VE153',
-21 silly get     vary: 'Accept-Encoding' } ]
-22 silly fetchPackageMetaData Error: Registry returned 404 for GET on https://registry.npmjs.org/uglify-css
-22 silly fetchPackageMetaData     at makeError (/home/s.yamagishi/.nvm/versions/node/v6.1.0/lib/node_modules/npm/node_modules/npm-registry-client/lib/request.js:264:12)
-22 silly fetchPackageMetaData     at CachingRegistryClient.<anonymous> (/home/s.yamagishi/.nvm/versions/node/v6.1.0/lib/node_modules/npm/node_modules/npm-registry-client/lib/request.js:242:14)
-22 silly fetchPackageMetaData     at Request._callback (/home/s.yamagishi/.nvm/versions/node/v6.1.0/lib/node_modules/npm/node_modules/npm-registry-client/lib/request.js:172:14)
-22 silly fetchPackageMetaData     at Request.self.callback (/home/s.yamagishi/.nvm/versions/node/v6.1.0/lib/node_modules/npm/node_modules/request/request.js:199:22)
-22 silly fetchPackageMetaData     at emitTwo (events.js:106:13)
-22 silly fetchPackageMetaData     at Request.emit (events.js:191:7)
-22 silly fetchPackageMetaData     at Request.<anonymous> (/home/s.yamagishi/.nvm/versions/node/v6.1.0/lib/node_modules/npm/node_modules/request/request.js:1036:10)
-22 silly fetchPackageMetaData     at emitOne (events.js:101:20)
-22 silly fetchPackageMetaData     at Request.emit (events.js:188:7)
-22 silly fetchPackageMetaData     at IncomingMessage.<anonymous> (/home/s.yamagishi/.nvm/versions/node/v6.1.0/lib/node_modules/npm/node_modules/request/request.js:963:12)
-22 silly fetchPackageMetaData  error for uglify-css { Error: Registry returned 404 for GET on https://registry.npmjs.org/uglify-css
-22 silly fetchPackageMetaData     at makeError (/home/s.yamagishi/.nvm/versions/node/v6.1.0/lib/node_modules/npm/node_modules/npm-registry-client/lib/request.js:264:12)
-22 silly fetchPackageMetaData     at CachingRegistryClient.<anonymous> (/home/s.yamagishi/.nvm/versions/node/v6.1.0/lib/node_modules/npm/node_modules/npm-registry-client/lib/request.js:242:14)
-22 silly fetchPackageMetaData     at Request._callback (/home/s.yamagishi/.nvm/versions/node/v6.1.0/lib/node_modules/npm/node_modules/npm-registry-client/lib/request.js:172:14)
-22 silly fetchPackageMetaData     at Request.self.callback (/home/s.yamagishi/.nvm/versions/node/v6.1.0/lib/node_modules/npm/node_modules/request/request.js:199:22)
-22 silly fetchPackageMetaData     at emitTwo (events.js:106:13)
-22 silly fetchPackageMetaData     at Request.emit (events.js:191:7)
-22 silly fetchPackageMetaData     at Request.<anonymous> (/home/s.yamagishi/.nvm/versions/node/v6.1.0/lib/node_modules/npm/node_modules/request/request.js:1036:10)
-22 silly fetchPackageMetaData     at emitOne (events.js:101:20)
-22 silly fetchPackageMetaData     at Request.emit (events.js:188:7)
-22 silly fetchPackageMetaData     at IncomingMessage.<anonymous> (/home/s.yamagishi/.nvm/versions/node/v6.1.0/lib/node_modules/npm/node_modules/request/request.js:963:12) pkgid: 'uglify-css', statusCode: 404, code: 'E404' }
-23 silly rollbackFailedOptional Starting
-24 silly rollbackFailedOptional Finishing
-25 silly runTopLevelLifecycles Starting
-26 silly runTopLevelLifecycles Finishing
-27 silly install printInstalled
-28 verbose stack Error: Registry returned 404 for GET on https://registry.npmjs.org/uglify-css
-28 verbose stack     at makeError (/home/s.yamagishi/.nvm/versions/node/v6.1.0/lib/node_modules/npm/node_modules/npm-registry-client/lib/request.js:264:12)
-28 verbose stack     at CachingRegistryClient.<anonymous> (/home/s.yamagishi/.nvm/versions/node/v6.1.0/lib/node_modules/npm/node_modules/npm-registry-client/lib/request.js:242:14)
-28 verbose stack     at Request._callback (/home/s.yamagishi/.nvm/versions/node/v6.1.0/lib/node_modules/npm/node_modules/npm-registry-client/lib/request.js:172:14)
-28 verbose stack     at Request.self.callback (/home/s.yamagishi/.nvm/versions/node/v6.1.0/lib/node_modules/npm/node_modules/request/request.js:199:22)
-28 verbose stack     at emitTwo (events.js:106:13)
-28 verbose stack     at Request.emit (events.js:191:7)
-28 verbose stack     at Request.<anonymous> (/home/s.yamagishi/.nvm/versions/node/v6.1.0/lib/node_modules/npm/node_modules/request/request.js:1036:10)
-28 verbose stack     at emitOne (events.js:101:20)
-28 verbose stack     at Request.emit (events.js:188:7)
-28 verbose stack     at IncomingMessage.<anonymous> (/home/s.yamagishi/.nvm/versions/node/v6.1.0/lib/node_modules/npm/node_modules/request/request.js:963:12)
-29 verbose statusCode 404
-30 verbose pkgid uglify-css
-31 verbose cwd /home/s.yamagishi/404/404/archive/v5
-32 error Linux 2.6.32-279.el6.x86_64
-33 error argv "/home/s.yamagishi/.nvm/versions/node/v6.1.0/bin/node" "/home/s.yamagishi/.nvm/versions/node/v6.1.0/bin/npm" "install" "uglify-css" "--save-dev"
-34 error node v6.1.0
-35 error npm  v3.8.6
-36 error code E404
-37 error 404 Registry returned 404 for GET on https://registry.npmjs.org/uglify-css
-38 error 404
-39 error 404 'uglify-css' is not in the npm registry.
-40 error 404 You should bug the author to publish it (or use the name yourself!)
-41 error 404 Note that you can also install from a
-42 error 404 tarball, folder, http url, or git url.
-43 verbose exit [ 1, true ]
