+++ archive/v5/package.json
@@ -9,8 +9,6 @@
     "underscore": "^1.8.3"
   },
   "devDependencies": {
-    "babel-preset-es2015": "^6.6.0",
-    "babelify": "^7.3.0",
     "browserify": "^13.0.1",
     "browserify-shim": "^3.8.12",
     "nodemon": "^1.9.2",
